fix: resolve workout save and coach data visibility issues
- Remove non-existent columns (template_id, variant, is_assisted, etc.) from
  exercise/set inserts that were silently failing, preventing data from saving
- Rewrite getRecentLifts to use separate queries instead of nested select to
  avoid RLS issues with PostgREST embedding
- Add decimal weight support (inputMode="decimal", parseFloat)
- Add L/R set numbering (1L, 1R, 2L, 2R)
- Make exercise library collapsible with scroll cap
- Persist log page navigation state via sessionStorage
- Restore in-progress workout from localStorage on navigation

+++ supabase/schema.sql
@@ -45,7 +45,7 @@ create table public.exercises (
 create table public.sets (
   id uuid default uuid_generate_v4() primary key,
   exercise_id uuid references public.exercises on delete cascade not null,
-  weight integer not null,
+  weight numeric(7,1) not null,
   reps integer not null,
   order_index integer not null default 0,
   created_at timestamptz default now() not null
